+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed MSRV to 1.74.1.
 - Added support for `@query` RFC-9421 component.
 - Exit with non-zero code if validation error occurs in `add-emoji` or `import-emoji` command.
+- Always validate actor profile data when it is retrieved from the database.
 
 ### Removed